Guidance simple print
This code is already populated with 42 "list of trees" that were crafted to advocate for veganism, aka to catch (42) common statements/misconceptions made and output the relevant reply. Both the number and nature of the output and input recognized can be altered in any way as long as the input list of trees follow specific rules (commented at the top of the code).